refactor: update canton modules structure
